+++ src/Components/WorkSpace/JoinSuccess/JoinSuccess.tsx
@@ -1,49 +1,19 @@
-import { useState, useEffect } from "react";
-import { SeugiCustomAxios } from '@/Api/SeugiCutomAxios';
+import React from "react";
 import Button from "@/Components/Button/Button";
 import * as S from "./JoinSuccess.style";
-import { useNavigate, useLocation } from "react-router-dom";
 import Backimg from "@/assets/image/Backimg.svg";
-// import config from '@/constants/config/config.json';
-import React from "react";
+
+import useJoinSuccess from '@/Hooks/JoinSuccess/index';
 
 const JoinSuccess = () => {
-  const navigate = useNavigate();
-  const [workspaceName, setWorkspaceName] = useState("");
-  const [schoolInfo, setSchoolInfo] = useState("");
-  const [schoolImgUrl, setSchoolImgUrl] = useState("");
-  const [workspaceId, setWorkspaceId] = useState("");
-  const location = useLocation();
-  const { verificationCode } = location.state || {};
-
-  const handleJoinSuccess = () => {
-    navigate('/selectjob', { state: { verificationCode, workspaceId } });
-  };
-
-  useEffect(() => {
-    const handleSchoolInfo = async () => {
-      try {
-
-        const res = await SeugiCustomAxios.get(`/workspace/search/${verificationCode}`);
-        const data = res.data.data;
-
-        setWorkspaceName(data.workspaceName);
-        setSchoolInfo(
-          `학생 ${data.studentCount}명 선생님 ${data.teacherCount}명`
-        );
-        setSchoolImgUrl(data.workspaceImageUrl);
-        setWorkspaceId(data.workspaceId);
-      } catch (error) {
-        console.error("Failed to fetch school information:", error);
-      }
-    };
-    handleSchoolInfo();
-  }, [verificationCode]);
-
-  const Backclick = () => {
-    navigate("/schoolcode");
-  };
-
+  const {
+    workspaceName,
+    schoolInfo,
+    schoolImgUrl,
+    handleJoinSuccess,
+    Backclick
+  } = useJoinSuccess();
+  
   return (
     <S.SuccessMain>
       <S.Container>
